Refactor: Outsource author event handler to its own module
- Moved author logic into a dedicated file (author.js)
- Added full JSDoc documentation for clarity and maintainability
- Replaced var with const for modern scoping
