+++ js/buttons.js
@@ -196,110 +196,6 @@ $(document).ready(function () {
     }
   });
 
-  /**
-   * Sets up the toggle functionality for contact person fields in author rows.
-   * When a contact person checkbox is checked, additional input fields for email and website
-   * are shown. When unchecked, these fields are hidden and cleared.
-   */
-  function setupContactPersonToggle() {
-    $("[data-creator-row]").each(function () {
-      var row = $(this);
-      var checkbox = row.find("[id^='checkbox-author-contactperson']");
-      var contactFields = row.find(".contact-person-input");
-
-      // Remove existing click handler to prevent duplicate bindings
-      checkbox.off("click");
-
-      function updateFields() {
-        if (checkbox.prop('checked')) {
-          contactFields.show();
-        } else {
-          contactFields.hide().find("input").val(""); // Clear input values when hiding
-        }
-      }
-
-      updateFields(); // Set initial state
-      checkbox.on("click", updateFields);
-    });
-  }
-
-  // Initial setup of contact person toggle functionality
-  setupContactPersonToggle();
-
-  /**
-  * Initialize sortable functionality for author rows
-  * Allows drag and drop reordering of authors using the drag handle
-  */
-  $("#group-author").sortable({
-    items: "[data-creator-row]",
-    handle: ".drag-handle",
-    axis: "y",
-    tolerance: "pointer",
-    containment: "parent"
-  });
-
-  // Store a clone of the original author row for later use
-  const originalAuthorRow = $("#group-author").children().first().clone();
-
-  /**
-  * Handles the addition of new author rows when the add button is clicked
-  * Creates a new row with unique IDs and proper event handlers
-  */
-  $("#button-author-add").click(function () {
-    var authorGroup = $("#group-author");
-    var newAuthorRow = originalAuthorRow.clone();
-
-    // Reset validation states and clear input values
-    newAuthorRow.find("input").val("").removeClass("is-invalid is-valid");
-    newAuthorRow.find(".invalid-feedback, .valid-feedback").css("display", "");
-
-    // Generate unique IDs for the new row's elements using timestamp
-    var uniqueSuffix = new Date().getTime();
-
-    // Update IDs of all relevant input fields with unique suffix
-    const fieldsToUpdate = [
-      "input-author-affiliation",
-      "input-author-rorid",
-      "input-contactperson-email",
-      "input-contactperson-website",
-      "checkbox-author-contactperson"
-    ];
-
-    fieldsToUpdate.forEach(fieldId => {
-      newAuthorRow.find(`#${fieldId}`).attr("id", `${fieldId}-${uniqueSuffix}`);
-    });
-
-    // Update label's 'for' attribute to match new checkbox ID
-    newAuthorRow.find("label.btn").attr("for", `checkbox-author-contactperson-${uniqueSuffix}`);
-
-    // Clean up and prepare new row
-    newAuthorRow.find(".tagify").remove();
-    newAuthorRow.find(".addAuthor").replaceWith(removeButton);
-    replaceHelpButtonInClonedRows(newAuthorRow);
-
-    // Add new row to the author group
-    authorGroup.append(newAuthorRow);
-
-    // Initialize autocomplete for affiliation field
-    autocompleteAffiliations(
-      `input-author-affiliation-${uniqueSuffix}`,
-      `input-author-rorid-${uniqueSuffix}`,
-      affiliationsData
-    );
-
-    // Add remove button functionality
-    newAuthorRow.on("click", ".removeButton", function () {
-      $(this).closest(".row").remove();
-    });
-
-    // Initialize Bootstrap tooltips
-    newAuthorRow.find('[data-bs-toggle="tooltip"]').each(function () {
-      const tooltip = new bootstrap.Tooltip(this);
-    });
-
-    // Reinitialize contact person toggle functionality for all rows
-    setupContactPersonToggle();
-  });
 
 
   /**
